% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/avg_age.R
\name{avg_age}
\alias{avg_age}
\title{An Average Age function}
\usage{
avg_age(year)
}
\arguments{
\item{year}{Enter the year you are interested in}
}
\description{
This function allows you to find out the average age of players in a specific year
}
\examples{
avg_age()
}
\keyword{avg_age}
