% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_matrix.R
\name{cor_matrix}
\alias{cor_matrix}
\title{Correlation Matrix}
\usage{
cor_matrix(year)
}
\arguments{
\item{year}{Enter the year you are interested in}
}
\description{
This function allows you to print the correlation matrix of numeric values in the NBA dataset
}
\examples{
cor_matrix()
}
\keyword{correlation}
