% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cormatrix_numeric.R
\name{cormatrix_numeric}
\alias{cormatrix_numeric}
\title{A Correlation Matrix Function}
\usage{
cormatrix_numeric(year)
}
\description{
This function allows you to include numeric continuous variables only for a certain year in the dataset, and generate a correlation matrix for all the numeric variables
}
\examples{
cormatrix_numeric(1995)
}
\keyword{correlation}
\keyword{matrix}
\keyword{numeric,}
