% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function1.R
\name{Function1}
\alias{Function1}
\title{A NBA Function 1}
\usage{
Function1()
This function allows you to find the Average Points scored in a specific Year.
}
\arguments{
\item{year}{Which specific year are you looking for? The dataset covers all seasons from 1950 to 2017.}
}
\description{
A NBA Function 1
}
\keyword{NBA}
