% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function2.R
\name{Function2}
\alias{Function2}
\title{A NBA Function 2}
\usage{
Function2()
This function allows you to find the oldest player in the NBA data for a specific Year.
}
\arguments{
\item{year}{Which specific year are you looking for? The dataset covers all seasons from 1950 to 2017.}
}
\description{
A NBA Function 2
}
\examples{
Function2()
}
\keyword{NBA}
