% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Function3.R
\name{Function3}
\alias{Function3}
\title{A NBA Function 3}
\usage{
Function3()
This function should limit the data for a specific year to include numeric continuous variables only (not categorical data) and should have the correlation matrix for all numeric variables in the data for that year.
}
\arguments{
\item{year}{Which specific year are you looking for? The dataset covers all seasons from 1950 to 2017.}
}
\description{
A NBA Function 3
}
\examples{
Function3()
}
\keyword{NBA}
