% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{A Age Function}
\usage{
oldest_player(year)
}
\description{
This function allows you to find the oldest player in a given year of the dataset.
}
\examples{
oldest_player(1995)
}
\keyword{age}
