% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_most.R
\name{score_most}
\alias{score_most}
\title{A Score Function}
\usage{
score_most(year)
}
\description{
This function allows you to find the player score the most points in a given year
}
\examples{
score_most(1995)
}
\keyword{points}
