% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oldest_player.R
\name{oldest_player}
\alias{oldest_player}
\title{Oldest Player}
\usage{
oldest_player(year)
}
\arguments{
\item{year}{Enter the year you are interested in}
}
\description{
This function allows you to find out the oldest player in a specific year
}
\examples{
oldest_player()
}
\keyword{oldest_player}
